feat. cart function

+++ src/main/java/com/socialcommerce/domain/cart/dto/CartResponse.java
@@ -1,5 +1,6 @@
 package com.socialcommerce.domain.cart.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -16,5 +17,5 @@ import lombok.experimental.FieldDefaults;
 public class CartResponse {
   private Long cartId;
   private List<CartItemResponse> cartItems;
-  private int totalPrice;
+  private BigDecimal totalPrice;
 }
